继续！！！

+++ ly-common/src/main/java/com/leyou/pojo/PageResult.java
@@ -3,37 +3,37 @@ package com.leyou.pojo;
 import java.util.List;
 
 public class PageResult<T> {
-    private Integer total;// 总条数
-    private Integer totalPage;// 总页数
+    private Long total;// 总条数
+    private Long totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult() {
     }
 
-    public PageResult(Integer total, List<T> items) {
+    public PageResult(Long total, List<T> items) {
         this.total = total;
         this.items = items;
     }
 
-    public PageResult(Integer total, Integer totalPage, List<T> items) {
+    public PageResult(Long total, Long totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
 
-    public Integer getTotalPage() {
+    public Long getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Integer totalPage) {
+    public void setTotalPage(Long totalPage) {
         this.totalPage = totalPage;
     }
